Move PerRequestSequelCache to separate file #2960
